# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NCDMB API Documentation",
        "description": "Comprehensive API documentation for the NCDMB Document Management System. This API provides endpoints for document management, workflow automation, procurement, legal processes, and more.",
        "contact": {
            "email": "support@ncdmb.gov.ng"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/legal/compliance-checks": {
            "get": {
                "tags": [
                    "Legal"
                ],
                "summary": "List all legal compliance checks",
                "operationId": "851a333e238b9ef76f12fef5cc62f6ef",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Legal"
                ],
                "summary": "Create a new legal compliance check",
                "operationId": "e0b427e2bfa5097b8ea52b95ce761de7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LegalComplianceCheckRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compliance check created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/legal/compliance-checks/{id}": {
            "get": {
                "tags": [
                    "Legal"
                ],
                "summary": "Get a specific legal compliance check",
                "operationId": "0b2001c86c614217ca868f780f6ba1d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Legal"
                ],
                "summary": "Update a legal compliance check",
                "operationId": "4aa8151c44342f3cbe4127534d28338e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LegalComplianceCheckRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compliance check updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Legal"
                ],
                "summary": "Delete a legal compliance check",
                "operationId": "071abe83c7024b2e31fd52e0a0dcfb8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Compliance check deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/legal/compliance-checks/contract/{contractId}": {
            "get": {
                "tags": [
                    "Legal"
                ],
                "summary": "Get compliance checks for a specific contract",
                "operationId": "ecff07b9ca4d1d313f3481b23be47abd",
                "parameters": [
                    {
                        "name": "contractId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Success"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Success": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "type": "object"
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation successful"
                    }
                },
                "type": "object"
            },
...
            "LegalComplianceCheckRequest": {
                "required": [
                    "project_contract_id",
                    "compliance_type",
                    "checked_by",
                    "check_date"
                ],
                "properties": {
                    "project_contract_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "compliance_type": {
                        "type": "string",
                        "enum": [
                            "procurement_act",
                            "fiscal_responsibility",
                            "public_accounts",
                            "company_law",
                            "tax_compliance",
                            "other"
                        ],
                        "example": "procurement_act"
                    },
                    "check_status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "passed",
                            "failed",
                            "conditional"
                        ],
                        "example": "pending",
                        "nullable": true
                    },
                    "checked_by": {
                        "type": "integer",
                        "example": 1
                    },
                    "check_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "findings": {
                        "type": "string",
                        "example": "Compliance findings",
                        "nullable": true
                    },
                    "corrective_actions": {
                        "type": "string",
                        "example": "Required corrective actions",
                        "nullable": true
                    },
                    "follow_up_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-15",
                        "nullable": true
                    },
                    "compliance_score": {
                        "type": "number",
                        "maximum": 100,
                        "minimum": 0,
                        "example": 90,
                        "nullable": true
                    },
                    "requires_remediation": {
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    },
                    "remediation_plan": {
                        "type": "string",
                        "example": "Remediation plan details",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token Authentication. Enter token in format: Bearer <token>",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}